fix(config-policy): break configurationPolicy↔policyBaselineDefaults import cycle (#1725)
The single-source refactor made pamSettings → policyBaselineDefaults →
configurationPolicy, transitively dragging the heavy configurationPolicy service
(which references db/schema tables in FEATURE_TABLE_MAP at module-eval) into the
agent helpers.*.test.ts suites. Those tests use a partial vi.mock('../../db/schema'),
so configurationPolicy crash-loaded there → 3 suites collected 0 tests → Test API red.

Extract CONFIG_FEATURE_TYPES / ConfigFeatureType into a leaf module
(configFeatureTypes.ts); configurationPolicy re-exports them for back-compat and
policyBaselineDefaults imports from the leaf instead. This removes the
policyBaselineDefaults → configurationPolicy edge, so pamSettings/helpers no
longer load the service. Verified the 3 previously-failing suites now pass.
